+++ Python/NeonOcean.S4.Main/NeonOcean/S4/Main/ThisNamespace.py
@@ -1,0 +1,1 @@
+Namespace = "NeonOcean.S4.Main"  # type: str